requirements.txt
Flask
flask-cors
pandas
numpy
scikit-learn
matplotlib
openpyxl
pillow 
pygame 
mysql-connector-python
requests
tkinter
difflib
sklearn.linear_model
sklearn.cluster
sklearn.tree
sklearn.svm
sklearn.neural_network
sklearn.metrics